030-最小的K个数: v1: feat: add InsertToSortedSlice()
切片的指针进行切片要使用括号包裹：(*buf)[i:]。因为

p[low:high] is shorthand for (*p)[low:high], that is, the pointer is dereferenced automatically for you. This does not happen automatically if p is a pointer to a slice, p[low:high] is invalid as you cannot slice a pointer. So you have to dereference the pointer manually in case of pointers to slices.

参考 https://stackoverflow.com/a/38014097/5238892